cli: Fix package import paths after renaming cli-go to cli #465
