+++ cli/zally/app_test.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/urfave/cli"
-	"github.com/zalando-incubator/zally/cli-go/zally/tests"
+	"github.com/zalando-incubator/zally/cli/zally/tests"
 )
 
 func TestApp(t *testing.T) {
