+++ cli/zally/commands/supported_rules.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 
 	"github.com/urfave/cli"
-	"github.com/zalando-incubator/zally/cli-go/zally/domain"
-	"github.com/zalando-incubator/zally/cli-go/zally/utils"
+	"github.com/zalando-incubator/zally/cli/zally/domain"
+	"github.com/zalando-incubator/zally/cli/zally/utils"
 )
 
 // SupportedRulesCommand lists supported rules
