+++ cli/zally/commands/supported_rules_test.go
@@ -14,9 +14,9 @@ import (
 	"time"
 
 	"github.com/urfave/cli"
-	"github.com/zalando-incubator/zally/cli-go/zally/domain"
-	"github.com/zalando-incubator/zally/cli-go/zally/tests"
-	"github.com/zalando-incubator/zally/cli-go/zally/utils"
+	"github.com/zalando-incubator/zally/cli/zally/domain"
+	"github.com/zalando-incubator/zally/cli/zally/tests"
+	"github.com/zalando-incubator/zally/cli/zally/utils"
 )
 
 func TestListRules(t *testing.T) {
