+++ cli/zally/readers/json_reader_test.go
@@ -3,7 +3,7 @@ package readers
 import (
 	"testing"
 
-	"github.com/zalando-incubator/zally/cli-go/zally/tests"
+	"github.com/zalando-incubator/zally/cli/zally/tests"
 
 	"encoding/json"
 )
